scrarping/requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
Django==4.2.11
django-cors-headers==4.3.1
djangorestframework==3.15.1
h11==0.14.0
idna==3.7
outcome==1.3.0.post0
Pygments==2.18.0
PySocks==1.7.1
selenium==4.20.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.0
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==2.2.1
wsproto==1.2.0
